MCP/requirements.txt
arxiv
python-dotenv
openai
notebook
mcp